Completed Refactor
