+++ resolver/resolver.go
@@ -1,4 +1,4 @@
-package rmapp
+package resolver
 
 /*
 Resolver.go holds the logic for the resolver which contains the associated
@@ -12,6 +12,9 @@ import (
 	"strings"
 
 	"github.com/alewtschuk/pfmt"
+	"github.com/alewtschuk/rmapp/deleter"
+	"github.com/alewtschuk/rmapp/finder"
+	"github.com/alewtschuk/rmapp/options"
 )
 
 // Resolver holds all the information reagarding the application's info
@@ -19,35 +22,28 @@ type Resolver struct {
 	AppName       string          // full .app name to be deleted
 	MdlsReturnStr string          // full return string of the mlds command call
 	BundleID      string          // app's bundle ID
-	Finder        Finder          // finder to look for files using app info
-	Options       ResolverOptions // resolver options
-	Deleter       Deleter         // deleter struct for handling file removal
+	Finder        finder.Finder   // finder to look for files using app info
+	Options       options.Options // resolver options
+	Deleter       deleter.Deleter // deleter struct for handling file removal
 	Peeked        bool            // resolved in peek mode
 }
 
-// Holds all command line related options
-type ResolverOptions struct {
-	Verbosity bool // is verbose flag set
-	Mode      bool // sets mode between trash and delete
-	Peek      bool // sets user peeking files to true
-}
-
 // Creates resolver struct and populates fields
-func NewResolver(app string, opts ResolverOptions) (*Resolver, bool) {
+func NewResolver(app string, opts options.Options) (*Resolver, bool) {
 	appName := getDotApp(app)
 	mdlsReturnStr := getMdlsIdentifier(appName)
 	if opts.Verbosity {
 		fmt.Println("Application to delete: ", pfmt.ApplyColor(app, 2))
 		fmt.Print("Resolved Bundle ID: ", pfmt.ApplyColor(getBundleID(mdlsReturnStr), 2), "\n\n")
 	}
-	finder, peeked := NewFinder(app, getBundleID((mdlsReturnStr)), opts) // uses app name over .app to ensure propper name based searching
+	finder, peeked := finder.NewFinder(app, getBundleID((mdlsReturnStr)), opts) // uses app name over .app to ensure propper name based searching
 	resolver := &Resolver{
 		AppName:       appName,
 		MdlsReturnStr: mdlsReturnStr,
 		BundleID:      getBundleID(mdlsReturnStr),
 		Finder:        finder,
 		Options:       opts,
-		Deleter:       NewDeleter(finder.MatchedPaths, opts),
+		Deleter:       deleter.NewDeleter(finder.MatchedPaths, opts),
 		Peeked:        peeked,
 	}
 	return resolver, peeked
