--- rmapp/report.go
@@ -1,1 +1,0 @@
-package rmapp
